change to use custom body reader
